Added proxy director and translated logs.
reverseProxy() func was using `httputil.NewSingleHostReverseProxy()` method and this method was directly appointing proxy director. Because of that, some path was repeated.

After many trying I tried to change the proxy director in the
reverseProxy() func and paths didn't repeat.

In addition to this, I translated log messages to English from Turkish.
Afterward, I converted the logging strategy more minimal and effective.