Added proxy director and translated logs.
As I said in this commit
https://github.com/samilboyalikli/learning-archive/commit/ae151cc95477c0291589efdf30b02c3adc209e39.
reverseProxy() func was using `httputil.NewSingleHostReverseProxy()`
method and this method was directly appointing proxy director. Because
of that, some path was repeated. And I wrote director manuelly. With
this way repeats stopped.

In addition to this, I translated log messages to English from Turkish.
Afterward, I converted the logging strategy more minimal and effective.