Big-o optimized and added a rate limiter func.
Datatype of ipStore was an IPData. But IPData structure was a slice. When
program try to find something big-o notation was being log(n). It's not
a problem for a small request traffic but when I use this demo for
gateway of a microservice structures, it'll be cause of latency.

So I changed datatype of ipStore. After this commit, ipStore will be
map. With this way, big-o notation will be log(1). It's reasonable
scale for a standart gateways.

I added a func named checkStore(). This func take an string ip and check ipStore. If string is exist in the store, increase count. Finally, if count values biggest than two, program is logging info. If biggest than three or equal, logging warning. With this way, we can see which ip's is cause of the traffic.