Coded a file for decimal/hexal/octal/binaries.
When i'm working on some problems in the platform. I tried to solve a
Python problem about converting number type of an input. This problem
was flagging as an easy but, I must admit that I find it so difficult.

This file takes an input and returns hexadecimal, octal and binary
equivalents of this decimal inputs and continues this process until it reaches zero.