Arithmetic operations func added to the file.
The question is user gives us two input and questions take them and
convert to integer. Afterward we're doing some arithmetic operations ot
them.

I wrote just a single line but my solutions was so primitive. I saw
another solution on this problem which are using array structure and *
on it. I didn't know sep parameter usage in the print lines with arrays.