The code blocks relocated.
The code block of the file relocated to be eclectic. Because, I wrote
the file directly without any relocation. Because of that, when I try to
add another block to the file, I had to scroll so down.

After this commit, I can write directly without any necessity to scroll.