Two solutions about list comprehension was written.
I saw a question in the Hackerrank about list comprehensions. I tried to
solve and I successed. But when I investigate my solution, found some
excesses.

Firstly, I used for keyword for checking condition of the list. But I
could check the condition inside of the list also.

So, I saw a solution by other coder. He/she wrote perfectly solution for
this problem. And he/she considered about scope of ability of the lists.