Some dictionary solutions.
There is a python question in Hackerrank. And solve this question almost
without any help. But I saw some unnecessary extra in my solution. I
used a for block and if block nested for. This code block solved the
problem perfectly with 4 lines.

But, I saw solution of the other coder. He/she just wrote 2 line without
any for or if blocks. So clear and genius solution. I liked it so much
and took the file.