Two solutions on a single list problem.
First solution from me. There is a group of inputs. And these inputs
include many commands. I took all these commands to a range and with if
block I can analyze all commands and I can apply these.

Second solution from other coder. Actually, I found my solution more
minimal than him/her. But, there is so beatiful sample of lambda funcs.
Because of that, I took this code block. He/she added a dictionary and
wrote commands as a key, wrote lambda funcs as a value. When take a
command, code block directly spliting that and appling on the
dictionary. With this way, commands is analyzing perfectly.