Solved a question about tuples.
I saw a question included tuples and hash function and I solved it. But
there is a important point: A coder asked about same solution is can not
working in the different version of python.

For example: This solutions which I wrote in the commit is working in the
pypy3 perfectly. But when you try to work this code block in python3 or
python2 will return error or wrong answer. After I saw this
question I searched that.

Actually this code block is working on Python 2. Just we must change
"input()" methods to the "raw_input()". But same hash() method will
give different results on pyhon2 and pypy3. Because hash() methods
are using different algorithms in python2 and pypy3.

I checked all hash algorithms with sys module in python. sys.hash_info()
func is giving us all configurations of hash function.