An example of using split() and join() functions.
I was mistake in the using of join() method. With this question I
learned true way. Lists must be used as a parameter in the join()
method. And the thing you want to add must be main object. I found it
interesting. Because, defaultly I was expecting the list is being the
main object in the usage of join method.