Solved 2 python problem.
I was working on a platform which give users some Python questions.
First problem was about creating string value with inputs. Second
problem about changing string inputs.

I saw usage of format() function in the first problem. I didn't use that
but my solution was true also. But as I see, I can use this function in
many scenario and this will be more confortable for me. As I remember
especially usage of format() common on logging codes.

I noticed that in the second problem: When I'm trying to solve that, I
was coding long ways. There is a short way, I can use list shortcuts.
For example this structure: '[:]' It's so usable. I saw other coders are
using that in discussion tab.