Solved a terrible Python question.
There was a question, asked from coders about count of a substring item
which exist in a string. But there is a problem: this string items
overlapping.

Unfortunately I couldn't find the solution and I took some help from AI.
After many trying, I found it and when I look at the discussion tab, I
saw our solution was best.