Sample usage of Common Table Expression and some funcs.
I solved a hard SQL question. Main request of question is a triagle but
made with stars. Star is string char like that '*'. But for solving
that, I must write a recursive code block and I didn't know how is it.

After searching some SQL keywords and functions. I found RECURSIVE
keyword. This keyword is doing recursion by changing the subject of the
function.

I learned some functions also as WITH and REPEAT(). With WITH, we can
create temporary tables and we can use it. The repeat function repeats
the first argument the same number of times as the second argument.