I wrote a function about some string methods.
There is some string methods which return boolean outputs to users as
`isalnum()`, `isalpha()`, `isdigit()`, `islower()` and `isupper()`.

Afterward I wrote a func for solving that. And when I look at the
discussion tab I saw the solution which coded by other coder and his/her
solutions is the better than mine.