Sample usage of map() method.
I solved a problem -actually I couldn't solve. How this problem
classified as easy I don't understand yet. This problem was incredible
hard to me.- This problem about recursive pattern. We created a pattern
with some string characters and we created a door map motif with it.

I learned usage of map() method in this problem, and when we are solving
that, we should be use some list shortcuts. It's was benefitial to me
also.