Alternative solution for previous python problem.
I saw another solution in the PyPy 3 tab. I surprized, because platform
gave us solution also -I'm trying to solve problems on Python 3-.

I learned that, when we are using join method with lists, we should use
concatenate lists too. Because join method can only concatenate list to
list.

Here we used pattern list in the join method with 'WELCOME' line and
added reverse version of pattern list to the end. With this way we can
print the requested result perfectly.