Some Backend Structures for Understanding BE layer.
I must learn backend technologies with go lang. Because of that, I'm
trying to code three simple backend sample. The first will be so
beginner level. The second will be standart backend and the third will
be so detailed and powerful. I hope, I can find time to coding for these three
backend samples.

I added main folders and go.mod file and go.sum file in this commit.